backend/requirements.txt
# Backend - Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

# Base de datos - MySQL / TiDB Cloud
PyMySQL==1.1.0
cryptography==41.0.7

# Utilidades
python-dotenv==1.0.0

# Servidor de producción
gunicorn==21.2.0
